Organize package for pip installation
