+++ scripts/clear_db.py
