+++ scripts/modelselector_eps_gridsearch_v2.py
@@ -5,7 +5,7 @@ import numpy as np
 import torch
 from tqdm import tqdm
 
-from datasets import Dataset
+from coda.datasets import Dataset
 from coda.baselines.modelpicker import ModelPicker
